refactor projectAnnotations import
