+++ packages/lib-classifier/src/plugins/tasks/survey/components/SurveyTask.userClicks.spec.js
@@ -2,14 +2,14 @@ import { expect } from 'chai'
 import { composeStory } from '@storybook/react'
 import { render, screen, waitFor, within } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import * as projectAnnotations from '../../../../../.storybook/preview'
+import projectAnnotations from '../../../../../../../.storybook/preview'
 
 import Meta, { Default } from './SurveyTask.stories'
 
 describe('SurveyTask with user clicks', function () {
   // this turns off Mocha's time limit for slow tests
   this.timeout(0)
-  const DefaultStory = composeStory(Default, Meta, projectAnnotations.default)
+  const DefaultStory = composeStory(Default, Meta, projectAnnotations)
 
   describe('when the Filter button is clicked', function () {
     let user, filterButton
